[feat]: small refactoring
