--- src/airflow/dag_tasks/__init__.py
@@ -1,13 +1,0 @@
-"""
-The dag_tasks package contains scripts for executing tasks in Airflow DAG.
-These scripts are intended to be run as individual tasks in the data processing pipeline.
-"""
-
-__all__ = [
-    'prepare_base_features',
-    'prepare_lstm_features',
-    'prepare_tft_features', 
-    'prepare_llm_features',
-    'init_feast',
-    'materialize_feast'
-]
