+++ src/airflow/dag_tasks/data_preparation/lstm/__init__.py
@@ -1,0 +1,10 @@
+"""
+The dag_tasks.data_preparation.lstm package contains data preparation scripts for LSTM model in Airflow DAG.
+These scripts are intended to be run as individual tasks in the data processing pipeline.
+"""
+
+__all__ = [
+    'prepare_lstm_features',
+    'encode_lstm_features',
+    'construct_lstm_dataset',
+]
