+++ src/airflow/dag_tasks/data_preparation/tft/__init__.py
@@ -1,0 +1,10 @@
+"""
+The dag_tasks.data_preparation.tft package contains data preparation scripts for TFT model in Airflow DAG.
+These scripts are intended to be run as individual tasks in the data processing pipeline.
+"""
+
+__all__ = [
+    'prepare_tft_features',
+    'encode_tft_features',
+    'construct_tft_dataset',
+]
