+++ src/airflow/dag_tasks/data_preparation/tft/encode_tft_features.py
@@ -10,10 +10,10 @@ import sys
 import logging
 import os
 
-sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
+sys.path.insert(0, str(Path(__file__).resolve().parents[5]))
 
-from src.utils import setup_logger
-from src.features.tft_feature_encoder import TFTFeatureEncoder
+from src.utils.data_utils import setup_logger
+from src.utils.data_utils.tft_feature_encoder import TFTFeatureEncoder
 from src.config.configs import settings
