+++ src/features/__init__.py
@@ -1,0 +1,12 @@
+"""
+The features package contains modules for feature engineering and data preparation.
+These modules are used to create features for different models in the pipeline.
+"""
+
+__all__ = [
+    'temporal_features',
+    'transaction_aggregation_features',
+    'llm_features',
+    'store_date_features',
+    'static_features',
+]
