+++ src/utils/__init__.py
@@ -1,24 +1,11 @@
-import sys
-from pathlib import Path
-import logging
+"""
+The utils package contains utility functions for the project.
+These functions are used to load and save data, setup the logger, scale features, encode features, and other utility functions.
+"""
 
-def find_project_root():
-    path = Path(__file__).resolve()
-    while not (path / 'pyproject.toml').exists():
-        path = path.parent
-        if path == path.parent:
-            break
-    return path
-
-
-sys.path.insert(0, str(find_project_root()))
-
-from src.utils.data_utils import (
-    setup_logger, 
-    load_data, 
-    save_data,
-    setup_seed
-)
-
-
-__all__ = ['setup_logger', 'load_data', 'save_data', 'setup_seed']
+__all__ = [
+    'data_utils',
+    'store_encoder',
+    'feature_scaler',
+    'tft_feature_encoder'
+]
